# api-contract/openapi.yaml
openapi: 3.0.3
info:
  title: Fullstack Workspace API
  version: 1.0.0
servers:
  - url: /api
    description: API base path
paths:
...
  /admin/roles:
    get:
      summary: 角色列表
      operationId: listRoles
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
        - name: pageSize
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - name: keyword
          in: query
          schema:
            type: string
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResultRoleList"
    post:
      summary: 创建角色
      operationId: createRole
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleCreate"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResultId"
  /admin/roles/{id}:
    get:
      summary: 角色详情
      operationId: getRole
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResultRoleDetail"
    put:
      summary: 更新角色
      operationId: updateRole
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleUpdate"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResultVoid"
    delete:
      summary: 删除角色
      operationId: deleteRole
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResultVoid"
  /admin/roles/{id}/permissions:
    get:
      summary: 角色权限列表
      operationId: getRolePermissions
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResultLongArray"
    put:
      summary: 分配角色权限
      operationId: assignRolePermissions
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RolePermissionIds"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResultVoid"
...
  /admin/menus:
    get:
      summary: 菜单树（管理用）
      operationId: getMenuTree
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResultMenuTree"
    post:
      summary: 创建菜单
      operationId: createMenu
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuCreate"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResultId"
  /admin/menus/{id}:
    put:
      summary: 更新菜单
      operationId: updateMenu
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuUpdate"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResultVoid"
    delete:
      summary: 删除菜单
      operationId: deleteMenu
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResultVoid"
  /admin/users:
    get:
      summary: 用户列表
      operationId: listAdminUsers
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
        - name: pageSize
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - name: keyword
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResultAdminUserList"
    post:
      summary: 创建用户
      operationId: createAdminUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserCreate"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResultId"
  /admin/users/{id}:
    get:
      summary: 用户详情
      operationId: getAdminUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResultAdminUserDetail"
    put:
      summary: 更新用户
      operationId: updateAdminUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResultVoid"
    delete:
      summary: 删除用户
      operationId: deleteAdminUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResultVoid"
  /admin/users/{id}/roles:
    get:
      summary: 用户角色列表
      operationId: getUserRoles
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResultLongArray"
    put:
      summary: 分配用户角色
      operationId: assignUserRoles
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRoleIds"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResultVoid"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResult:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: "0 成功, 1000 参数校验错误"
        message:
          type: string
        data:
          type: object
    ApiResultVoid:
      allOf:
        - $ref: "#/components/schemas/ApiResult"
        - properties:
            data:
              type: object
              nullable: true
...
    RoleListResponse:
      type: object
      properties:
        list:
          type: array
          items:
            $ref: "#/components/schemas/RoleListItem"
        total:
          type: integer
    RoleListItem:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        name:
          type: string
        description:
          type: string
        type:
          type: integer
        status:
          type: integer
        createTime:
          type: string
          format: date-time
    RoleDetail:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        name:
          type: string
        description:
          type: string
        type:
          type: integer
        status:
          type: integer
        createTime:
          type: string
          format: date-time
        updateTime:
          type: string
          format: date-time
    RoleCreate:
      type: object
      required:
        - code
        - name
        - type
      properties:
        code:
          type: string
        name:
          type: string
        description:
          type: string
        type:
          type: integer
        status:
          type: integer
    RoleUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: integer
    RolePermissionIds:
      type: object
      required:
        - permissionIds
      properties:
        permissionIds:
          type: array
          items:
            type: integer
...
    MenuManagementTreeItem:
      type: object
      properties:
        id:
          type: integer
        permissionId:
          type: integer
        parentId:
          type: integer
        name:
          type: string
        path:
          type: string
        component:
          type: string
        icon:
          type: string
        sortOrder:
          type: integer
        visible:
          type: integer
        children:
          type: array
          items:
            $ref: "#/components/schemas/MenuManagementTreeItem"
    MenuCreate:
      type: object
      required:
        - permissionId
        - parentId
        - name
      properties:
        permissionId:
          type: integer
        parentId:
          type: integer
        name:
          type: string
        path:
          type: string
        component:
          type: string
        icon:
          type: string
        sortOrder:
          type: integer
        visible:
          type: integer
    MenuUpdate:
      type: object
      properties:
        permissionId:
          type: integer
        parentId:
          type: integer
        name:
          type: string
        path:
          type: string
        component:
          type: string
        icon:
          type: string
        sortOrder:
          type: integer
        visible:
          type: integer
    AdminUserListResponse:
      type: object
      properties:
        list:
          type: array
          items:
            $ref: "#/components/schemas/AdminUserListItem"
        total:
          type: integer
    AdminUserListItem:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        nickname:
          type: string
        avatarUrl:
          type: string
        status:
          type: integer
        createTime:
          type: string
          format: date-time
    AdminUserDetail:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        nickname:
          type: string
        avatarUrl:
          type: string
        gender:
          type: integer
        birthday:
          type: string
        bio:
          type: string
        status:
          type: integer
        createTime:
          type: string
          format: date-time
        updateTime:
          type: string
          format: date-time
    AdminUserCreate:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        nickname:
          type: string
        status:
          type: integer
    AdminUserUpdate:
      type: object
      properties:
        nickname:
          type: string
        avatarUrl:
          type: string
        gender:
          type: integer
        birthday:
          type: string
        bio:
          type: string
        status:
          type: integer
    UserRoleIds:
      type: object
      required:
        - roleIds
      properties:
        roleIds:
          type: array
          items:
            type: integer
    IdResponse:
      type: object
      properties:
        id:
          type: integer
...
    ApiResultRoleList:
      allOf:
        - $ref: "#/components/schemas/ApiResult"
        - properties:
            data:
              $ref: "#/components/schemas/RoleListResponse"
    ApiResultRoleDetail:
      allOf:
        - $ref: "#/components/schemas/ApiResult"
        - properties:
            data:
              $ref: "#/components/schemas/RoleDetail"
    ApiResultId:
      allOf:
        - $ref: "#/components/schemas/ApiResult"
        - properties:
            data:
              $ref: "#/components/schemas/IdResponse"
    ApiResultLongArray:
      allOf:
        - $ref: "#/components/schemas/ApiResult"
        - properties:
            data:
              type: array
              items:
                type: integer
...
    ApiResultMenuTree:
      allOf:
        - $ref: "#/components/schemas/ApiResult"
        - properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/MenuManagementTreeItem"
    ApiResultAdminUserList:
      allOf:
        - $ref: "#/components/schemas/ApiResult"
        - properties:
            data:
              $ref: "#/components/schemas/AdminUserListResponse"
    ApiResultAdminUserDetail:
      allOf:
        - $ref: "#/components/schemas/ApiResult"
        - properties:
            data:
              $ref: "#/components/schemas/AdminUserDetail"